Modify pylint's module name regex
This makes it compatible with Django's naming convention for migration
modules, e.g. 0001_initial.py

As noted in the comment in pyproject.toml, it would be nice to allow
such names only in the migrations/ directory. But that's apparently not
supported. Admittedly, it's uncommon for people to try to put digits in
their module names, so this new regex is unlikely to cause problems.