Set queryset and initial value in __init__()
Unlike initial, queryset can't be assigned a callable when defining the
fields. So to use the new static methods, the queryset must be assigned
in __init__().

For the cateogry field, the initial value is now the first element of
the queryset, if there is one. This means we can get rid of the empty
label.