Update gitignore rules
As mentioned in the comment, package-lock.json is ignored because this
project only a single development dependency: eslint. Explicitly
tracking every indirect dependency (of which there are many) is overkill
and leads to unhelpful activity from GitHub's dependabot.

package-lock.json can be tracked in the future if runtime dependencies
are added.