Layout form horizontally
This involves rendering the form using <div> elements, as opposed to the
default <table>.

The built-in div.html template is overridden to set the class for
Bootstrap styling. Again, I'd like to extend or include the original
instead of copying the whole thing. But it's not set up to be
selectively tweaked like that.