Change row select mode
Now you can select multiple rows simply by clicking on them. Before, you
had to hold down Ctrl. The new way matches the way one selects multiple
emails in Gmail.

Selecting a range is a little more finicky: press and hold Shift, then
click and hold on a checkbox, then drag over the rows to select. If you
click next to the checkbox instead of exactly on it, it doesn't work
properly.

But I think the improved simplicity for selecting non-contiguous rows
outweighs the clunkiness of selecting a range. One can usually filter so
that only the desired rows are visible and then easily select them all
using the checkbox in the header.