Create new form for editing existing entries
The original EntryForm (now renamed to CreateEntryForm) could be used to
edit existing entries if the entry ID was already known to the server
(usually from the URL).

But now we want to be able to specify the ID within the form itself. The
cleanest way to do that is a new form with a new field.