Interpret updates value as dict, not list of dicts
This was due to a misunderstanding on my part about the way the request
data was encapsulated. When logging the data, it looks like each value
of a QueryDict is a list:
    <QueryDict: 'updates': ['{"deletions":[],"edits":[]}']}>
However, when you access the value, it returns the enclosed element, not
a list.

QueryDict is trying to be helpful since a field could be given multiple
times (hence they're stored internally as a list), but when accessing it
gives you the final value, not all of them.