Show footer without needing to scroll down
By moving the header within the top-level container, its height is now
properly accounted for. Previously, the header plus a separate 100% VH
container caused the footer to appear off the bottom of the page, even
if there was enough space to show it. Now the header's height counts
towards the container's 100% VH, so we don't add excess blank space
between the table and footer.