Drop empty tags when sending updates
When a user clears the Tags cell for an existing entry in the table, the
tag is set to the empty string on the client side. We don't want to send
this to the server since that would associate the entry with a tag of
"". Instead, we want to drop empty strings so the entry is not
associated with any tags.